# webdoc/public/openapi/queen-broker.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Queen MQ broker API",
    "version": "1.0.0",
    "summary": "The HTTP API of the Queen MQ broker.",
    "description": "Generated from the broker's own source at documentation build time by\n`webdoc/scripts/gen-openapi.mjs`: paths and methods come from the axum router,\nquery parameters from each handler's extractor, request bodies from the structs\nhandlers deserialize into, response codes from the `StatusCode` variants each\nhandler can return, and authorization from `auth::route_access_level`.\n\nResponse bodies on the hot paths are assembled by string concatenation rather than\nserialized from a type, and several management routes forward a stored procedure's\nJSON verbatim. Those responses carry `x-queen-schema: \"opaque\"` and are typed as an\nopen object: the field-by-field shape is documented on the reference pages instead of\nguessed at here.\n\nEvery 204 response has no body at all, not an empty JSON document.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6632",
      "description": "A broker with default settings"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Only enforced when JWT_ENABLED is on. Access levels are a role set, not a ladder: a write-only credential passes POST /api/v1/push and is rejected everywhere else."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "description": "The broker's error body."
      }
    }
  },
  "paths": {
...
    "/api/v1/configure": {
      "post": {
        "operationId": "configure",
        "tags": [
          "Queues"
        ],
        "summary": "POST /api/v1/configure",
        "parameters": [
          {
            "name": "x-queen-tenant",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tenant to scope this request to. Read only when QUEEN_TENANCY_HEADER is on; unauthenticated by design, so the broker must not be reachable directly by tenants."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape not derivable from a Rust type. See the route's page under /reference/http for the field-by-field contract."
                }
              }
            },
            "x-queen-schema": "opaque"
          },
          "400": {
            "description": "Malformed request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-queen-access-level": "read-write",
        "x-queen-tenant-scoped": true,
        "x-queen-handler": "server/src/handlers/queues.rs::handle_configure",
        "x-queen-query-params": "none",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "queue": {
                    "description": "See the reference page for this route."
                  },
                  "options": {
                    "description": "See the reference page for this route."
                  },
                  "error": {
                    "description": "See the reference page for this route."
                  }
                },
                "additionalProperties": true
              }
            }
          },
          "x-queen-schema": "opaque",
          "description": "The handler reads this body as a dynamic JSON document, so the exact shape is not recoverable from a type. The listed keys are the ones the handler reads by name."
        }
      }
    },
...
    "/api/v1/resources/queues/{queue}": {
      "get": {
        "operationId": "get_queue",
        "tags": [
          "Queues"
        ],
        "summary": "GET /api/v1/resources/queues/:queue",
        "parameters": [
          {
            "name": "queue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-queen-tenant",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tenant to scope this request to. Read only when QUEEN_TENANCY_HEADER is on; unauthenticated by design, so the broker must not be reachable directly by tenants."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape not derivable from a Rust type. See the route's page under /reference/http for the field-by-field contract."
                }
              }
            },
            "x-queen-schema": "opaque"
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-queen-access-level": "read-only",
        "x-queen-tenant-scoped": true,
        "x-queen-handler": "server/src/handlers/queues.rs::handle_get_queue",
        "x-queen-query-params": "none",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "delete_queue",
        "tags": [
          "Queues"
        ],
        "summary": "DELETE /api/v1/resources/queues/:queue",
        "parameters": [
          {
            "name": "queue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-queen-tenant",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Tenant to scope this request to. Read only when QUEEN_TENANCY_HEADER is on; unauthenticated by design, so the broker must not be reachable directly by tenants."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape not derivable from a Rust type. See the route's page under /reference/http for the field-by-field contract."
                }
              }
            },
            "x-queen-schema": "opaque"
          },
          "500": {
            "description": "Internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-queen-access-level": "admin",
        "x-queen-tenant-scoped": true,
        "x-queen-handler": "server/src/handlers/queues.rs::handle_delete_queue",
        "x-queen-query-params": "none",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}